backend/requirements.txt
Django==3.2.3
Pillow==9.0.0
djangorestframework==3.12.4
djoser==2.1.0
drf-extra-fields==3.7.0
gunicorn==20.1.0
django-filter==21.1.0
djangorestframework-filters